Re-use availability change behaviour explanation alert for the feature discovery
