+++ Cartfile.resolved
@@ -16,13 +16,13 @@ github "wireapp/protobuf-objc" "1.9.14"
 github "wireapp/swift-protobuf" "1.2.1"
 github "wireapp/wire-ios-canvas" "9.0.2"
 github "wireapp/wire-ios-cryptobox" "17.0.0"
-github "wireapp/wire-ios-data-model" "feature/alert-availability-on-upgrade"
+github "wireapp/wire-ios-data-model" "167.3.0"
 github "wireapp/wire-ios-images" "25.0.0"
 github "wireapp/wire-ios-link-preview" "21.0.0"
 github "wireapp/wire-ios-protos" "19.0.0"
 github "wireapp/wire-ios-request-strategy" "145.0.0"
 github "wireapp/wire-ios-share-engine" "133.0.0"
-github "wireapp/wire-ios-sync-engine" "fix/availability-notification"
+github "wireapp/wire-ios-sync-engine" "249.1.0"
 github "wireapp/wire-ios-system" "26.0.1"
 github "wireapp/wire-ios-testing" "18.0.0"
 github "wireapp/wire-ios-transport" "55.0.0"
