+++ Wire-iOS/Sources/UserInterface/Availability/AvailabilityTitleView.swift
@@ -175,28 +175,14 @@ class AvailabilityTitleView: TitleView, Themeable, ZMUserObserver {
             alert.addAction(UIAlertAction(title: availability.localizedName, style: .default, handler: { [weak self] (action) in
                 self?.didSelectAvailability(availability)
                 
-                if Settings.shared()?.shouldRemindUserWhenChanging(availability) == true, let informationAlert = self?.reminderSheet(for: availability) {
-                    presentingViewController.present(informationAlert, animated: true)
+                if Settings.shared()?.shouldRemindUserWhenChanging(availability) == true {
+                    presentingViewController.present(UIAlertController.availabilityExplanation(availability), animated: true)
                 }
             }))
         }
         
         alert.popoverPresentationController?.permittedArrowDirections = [ .up ]
         alert.addAction(UIAlertAction(title: "availability.message.cancel".localized, style: .cancel, handler: nil))
-        
-        return alert
-    }
-    
-    func reminderSheet(for availability: Availability) -> UIAlertController {
-        
-        let title = "availability.reminder.\(availability.canonicalName).title".localized
-        let message = "availability.reminder.\(availability.canonicalName).message".localized
-        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        
-        alert.addAction(UIAlertAction(title: "availability.reminder.action.dont_remind_me".localized, style: .default, handler: { (_) in
-            Settings.shared()?.dontRemindUserWhenChanging(availability)
-        }))
-        alert.addAction(UIAlertAction(title: "availability.reminder.action.ok".localized, style: .default, handler: { (_) in }))
         
         return alert
     }
